+++ src/Point.h
@@ -10,26 +10,26 @@
 // to a vertical axis coordinate.
 class Point {
 
-    std::tuple<float, float> XY;  // x-axis and y-axis coordinates, only accessed directly by private getter/setter
+    std::tuple<double, double> XY;  // x-axis and y-axis coordinates, only accessed directly by private getter/setter
 
-    std::tuple<float, float> getXY() { return Point::XY; }
-    void setXY(std::tuple<float, float> xy) { Point::XY = xy; }
+    std::tuple<double, double> getXY() { return Point::XY; }
+    void setXY(std::tuple<double, double> xy) { Point::XY = xy; }
 
 public:
 
     Point();
-    Point(float x, float y);
-    Point(std::tuple<float, float> xy);
+    Point(double x, double y);
+    Point(std::tuple<double, double> xy);
 
-    std::tuple<float, float> xy();
-    void xy(float *x, float *y);
-    float x();
-    float y();
+    std::tuple<double, double> xy();
+    void xy(double *x, double *y);
+    double x();
+    double y();
 
-    void xy(std::tuple<float, float> xy);
-    void xy(float x, float y);
-    void x(float x);
-    void y(float y);
+    void xy(std::tuple<double, double> xy);
+    void xy(double x, double y);
+    void x(double x);
+    void y(double y);
 
     virtual void print();
 };
